refactor: 모달의 다시 시작하기 버튼을 클릭하면 앱(Lotto)을 새로 마운트한다
